requirements.txt
transformers[torch]==4.16.2 
fastapi==0.75.0
requests==2.27.1
uvicorn[standard]==0.16.0
gunicorn==20.1.0
cloud-sql-python-connector[pg8000]==0.5.1
sqlmodel==0.0.6
sqlalchemy-utils==0.38.2
fastapi-auth0==0.3.0
python-dotenv==0.19.2 
loguru==0.6.0 
pyjwt[crypto]==2.3.0
fastapi-mail==1.0.6
